chore: upgrade to v7
