+++ package.json
@@ -26,8 +26,8 @@
     "format:write": "prettier --write packages/**/**/src --cache",
     "build:addon": "cd tooling/storybook-addon && pnpm build:fast",
     "clean:storybook": "rm -rf node_modules/.cache/storybook",
-    "storybook": "pnpm build:addon && start-storybook -p 6006",
-    "build:storybook": "pnpm build:addon && build-storybook && sb extract",
+    "storybook": "pnpm build:addon && storybook dev",
+    "build:storybook": "pnpm build:addon && storybook start && storybook extract",
     "serve:storybook": "pnpx serve storybook-static",
     "clean": "pnpm -r --parallel exec rimraf dist .turbo *.log",
     "changelog": "tsx scripts/changelog.ts",
@@ -64,14 +64,13 @@
     "@commitlint/cli": "14.1.0",
     "@commitlint/config-conventional": "14.1.0",
     "@octokit/rest": "^19.0.4",
-    "@storybook/addon-a11y": "^6.5.15",
-    "@storybook/addon-essentials": "^6.5.15",
-    "@storybook/addon-storysource": "^6.5.15",
-    "@storybook/builder-webpack5": "^6.5.15",
-    "@storybook/cli": "^6.5.15",
-    "@storybook/manager-webpack5": "^6.5.15",
-    "@storybook/react": "^6.5.15",
-    "@storybook/theming": "^6.5.15",
+    "@storybook/addon-a11y": "^7.0.18",
+    "@storybook/addon-essentials": "^7.0.18",
+    "@storybook/addon-storysource": "^7.0.18",
+    "@storybook/cli": "^7.0.18",
+    "@storybook/react": "^7.0.18",
+    "@storybook/react-vite": "^7.0.18",
+    "@storybook/theming": "^7.0.18",
     "@swc-node/jest": "^1.5.2",
     "@swc/core": "^1.2.177",
     "@testing-library/jest-dom": "^5.16.4",
@@ -97,16 +96,16 @@
     "eslint-plugin-react": "7.30.1",
     "eslint-plugin-react-hooks": "4.6.0",
     "eslint-plugin-testing-library": "5.5.1",
-    "fs-extra": "11.1.1",
     "find-packages": "9.0.13",
     "find-up": "^6.3.0",
+    "fs-extra": "11.1.1",
     "husky": "^7.0.4",
     "jest": "^28.1.1",
     "jest-environment-jsdom": "^28.1.1",
     "jest-watch-typeahead": "1.1.0",
     "lint-staged": "11.2.6",
-    "npm-run-all": "^4.1.5",
     "node-fetch": "3.3.1",
+    "npm-run-all": "^4.1.5",
     "ora": "6.1.2",
     "outdent": "^0.8.0",
     "plop": "3.1.2",
@@ -118,7 +117,9 @@
     "tsx": "^3.8.1",
     "turbo": "1.3.4",
     "typescript": "^5.0.2",
-    "unbuild": "0.7.4"
+    "unbuild": "0.7.4",
+    "vite": "^4.0.2",
+    "storybook": "^7.0.18"
   },
   "pnpm": {
     "overrides": {
