+++ packages/components/menu/stories/menu.stories.tsx
@@ -23,6 +23,19 @@ import {
   useMenuItem,
 } from "../src"
 
+export * from "./menu.stories"
+
+export default {
+  title: "Components / Overlay / Menu",
+  decorators: [
+    (story: Function) => (
+      <chakra.div maxWidth="500px" mx="auto" mt="40px">
+        {story()}
+      </chakra.div>
+    ),
+  ],
+}
+
 const words = [
   "About Visual Studio Code",
   "Check for updates",
